requirements.txt
streamlit==1.48.0
tensorflow-cpu==2.19.0
nltk==3.9.1
numpy==2.1.3
pandas==2.3.1
scikit-learn==1.5.2
pillow==11.3.0
matplotlib==3.10.5
